more debugging

+++ Nwpw/nwpwlib/device/nonlocal.cl
@@ -5,12 +5,12 @@ __kernel void Generate_projectors(const int ii, const int ng0, const int nprj, c
                                   const __global int    *indxi,
                                   const __global int    *indxj,
                                   const __global int    *indxk,
-                                  const __global double *phfacx,
-                                  const __global double *phfacy,
-                                  const __global double *phfacz,
+                                  const __global float *phfacx,
+                                  const __global float *phfacy,
+                                  const __global float *phfacz,
                                   const __global int    *sdfunction, 
-                                  const __global double *vnl,
-                                  __global double *prj) {
+                                  const __global float *vnl,
+                                  __global float *prj) {
    int shftx = 2*ii*nx;
    int shfty = 2*ii*ny;
    int shftz = 2*ii*ny;
@@ -19,13 +19,13 @@ __kernel void Generate_projectors(const int ii, const int ng0, const int nprj, c
    int i = get_global_id(0); //ng
    //int l = get_global_id(1); //nprj
    
-   double ai = phfacx[shftx+2*indxi[i]]; double bi = phfacx[shftx+2*indxi[i]+1];
-   double aj = phfacy[shfty+2*indxj[i]]; double bj = phfacy[shfty+2*indxj[i]+1];
-   double ak = phfacz[shftz+2*indxk[i]]; double bk = phfacz[shftz+2*indxk[i]+1];
-   double c  = aj*ak - bj*bk;
-   double d  = aj*bk + ak*bj;
-   double rexi = (ai*c - bi*d);
-   double iexi = (ai*d + bi*c);
+   float ai = phfacx[shftx+2*indxi[i]]; float bi = phfacx[shftx+2*indxi[i]+1];
+   float aj = phfacy[shfty+2*indxj[i]]; float bj = phfacy[shfty+2*indxj[i]+1];
+   float ak = phfacz[shftz+2*indxk[i]]; float bk = phfacz[shftz+2*indxk[i]+1];
+   float c  = aj*ak - bj*bk;
+   float d  = aj*bk + ak*bj;
+   float rexi = (ai*c - bi*d);
+   float iexi = (ai*d + bi*c);
 
    for (int l=0; l<nprj; ++l)
    {
